[3/3] Settings : Add user control over WiFi regulatory domain (country)
WiFi regulatory domain is determind based on the MCC of the currently connected mobile network.

This allows to change the current country setting (which will stick till the next mobile network change) or to set a constant override which will stick also after mobile network change.

Do note that this is not legal and you should check applicable local laws before using this !

Change-Id: I383442aec8d01953303370d5e065d8e80b920676
Signed-off-by: Jean-Pierre Rasquin <yank555.lu@gmail.com>

Conflicts:
	res/values/slim_arrays.xml